requirements.txt
beautifulsoup4==4.14.3
bleach==6.3.0
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
deep-translator==1.11.4
filelock==3.25.2
Flask==3.1.3
fonttools==4.62.1
fsspec==2026.2.0
idna==3.11
imbalanced-learn==0.14.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kaggle==2.0.0
kagglesdk==0.1.16
kiwisolver==1.5.0
langdetect==1.0.9
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
nltk==3.9.3
numpy==2.4.3
packaging==26.0
pandas==3.0.1
pillow==12.1.1
protobuf==7.34.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
regex==2026.2.28
requests==2.32.5
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
setuptools==82.0.1
six==1.17.0
sklearn-compat==0.1.5
soupsieve==2.8.3
sympy==1.14.0
text-unidecode==1.3
textblob==0.19.0
threadpoolctl==3.6.0
torch==2.10.0
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
webencodings==0.5.1
Werkzeug==3.1.6
